update RemoveModule [ClassController]

+++ Controllers/ClassController.cs
@@ -72,7 +72,7 @@ namespace kroniiapi.Controllers
 
             foreach (Class c in classList)
             {
-                c.Trainer = await _trainerService.GetTrainerById(c.TrainerId);
+                //c.Trainer = await _trainerService.GetTrainerById(c.TrainerId);
                 c.Admin = await _adminService.GetAdminById(c.AdminId);
             }
             IEnumerable<ClassResponse> classListDto = _mapper.Map<IEnumerable<ClassResponse>>(classList);
@@ -634,10 +634,7 @@ namespace kroniiapi.Controllers
             {
                 return NotFound(new ResponseDTO(404, "Class is not exist"));
             }
-            if(classInfor.TrainerId != assignModuleInput.TrainerId)
-            {
-                return BadRequest(new ResponseDTO(409, "Trainer are not in this class"));
-            }
+
             var moduleMarkState = await _markService.GetMarkByModuleId(assignModuleInput.ModuleId, null, null);
             if (moduleMarkState.Count() != 0)
             {
@@ -648,6 +645,12 @@ namespace kroniiapi.Controllers
             if (moduleCertificateState.Count() != 0)
             {
                 return NotFound(new ResponseDTO(404, "Trainee in this class has certificate with this module"));
+            }
+
+            var classModuleInfor = await _classService.GetClassModule(assignModuleInput.ClassId, assignModuleInput.ModuleId);
+            if(classModuleInfor.TrainerId != assignModuleInput.TrainerId)
+            {
+                return BadRequest(new ResponseDTO(409, "This trainer dont teach this class"));
             }
 
             int removeStatus = await _classService.RemoveModuleFromClass(assignModuleInput.ClassId, assignModuleInput.ModuleId);
